core: move the wasm-specific logging and input-method code out of line
Both were inlined into files that are otherwise platform-independent, behind
long #if defined(__EMSCRIPTEN__) blocks that dominated them.

The message throttle and console.* output move to core/application/WasmLogging;
SafeQApplication::DebugOutput now has one wasm branch calling into it. The IME
retargeting moves to score/graphics/WasmInputMethod, leaving GraphicsItem.cpp
with the non-wasm stubs (708 lines to 164).

The IME diagnostics that surrounded the retargeting code -- the DOM/window
snapshots, imeDescribe, the JS install and the logging flag -- go with them,
along with retargetInputMethod's context argument, which only fed those logs.
