+++ src/lib/core/application/SafeQApplication.cpp
@@ -18,9 +18,7 @@
 #include <wobjectimpl.h>
 
 #if defined(__EMSCRIPTEN__)
-#include <emscripten/console.h>
-
-#include <mutex>
+#include <core/application/WasmLogging.hpp>
 #endif
 
 W_OBJECT_IMPL(SafeQApplication)
@@ -28,82 +26,9 @@ W_OBJECT_IMPL(SafeQApplication)
 SafeQApplication::~SafeQApplication() { }
 
 #if defined(__EMSCRIPTEN__)
-namespace
-{
-// Repeated-message throttle, wasm only.
-//
-// On wasm every log line is an fprintf that goes through _fd_write into
-// console.*, and the browser captures a stack for each one. A runaway warning
-// -- Qt's own "QRhiGles2: Context is lost." once per frame, say, which we
-// cannot patch out -- therefore does not merely fill the console: it makes
-// DevTools unresponsive, which is exactly when the user needs it to run the
-// diagnostics. Collapse consecutive identical messages, syslog style.
-//
-// Deliberately not enabled on desktop: there a flood is survivable, stderr is
-// cheap, and dropping lines would change what developers see in a build they
-// rely on for debugging. Flipping that is a matter of widening this #if.
-constexpr qint64 log_throttle_threshold = 5;
-constexpr qint64 log_throttle_heartbeat = 10000;
-
-std::mutex g_throttleMutex;
-QString g_lastMessage;
-qint64 g_repeatCount = 0;
-thread_local bool t_lastSuppressed = false;
-
-// Emit one already-formatted line in a single call.
-//
-// stdio is pathologically expensive here: fprintf goes fiprintf -> vfiprintf ->
-// __stdio_write -> _fd_write -> doWritev -> write -> put_char, i.e. it crosses
-// into JS one character at a time, and the whole line is then re-parsed out of
-// the fd. emscripten_console_* hands the string straight to console.* instead.
-// Like fprintf, these never re-enter Qt's logging, so they are also safe to use
-// from inside the message handler.
-void emitLine(QtMsgType type, const char* line)
-{
-  switch(type)
-  {
-    case QtDebugMsg:
-    case QtInfoMsg:
-      emscripten_console_log(line);
-      break;
-    case QtWarningMsg:
-      emscripten_console_warn(line);
-      break;
-    case QtCriticalMsg:
-    case QtFatalMsg:
-      emscripten_console_error(line);
-      break;
-  }
-}
-
-// Returns how many suppressed repeats the caller should report now (0 for
-// none), and sets @p suppress when this message must be dropped.
-qint64 throttleStep(const QString& msg, bool& suppress)
-{
-  const std::lock_guard lock{g_throttleMutex};
-
-  if(msg == g_lastMessage)
-  {
-    ++g_repeatCount;
-    suppress = g_repeatCount > log_throttle_threshold;
-    // Still say something once in a while: a permanently repeating message
-    // must not look like silence.
-    if(suppress && (g_repeatCount % log_throttle_heartbeat) == 0)
-      return g_repeatCount;
-    return 0;
-  }
-
-  const qint64 pending = g_repeatCount > log_throttle_threshold ? g_repeatCount : 0;
-  g_lastMessage = msg;
-  g_repeatCount = 1;
-  suppress = false;
-  return pending;
-}
-}
-
 bool SafeQApplication::lastMessageWasSuppressed() noexcept
 {
-  return t_lastSuppressed;
+  return score::wasm::lastMessageWasSuppressed();
 }
 #else
 bool SafeQApplication::lastMessageWasSuppressed() noexcept
@@ -115,59 +40,21 @@ bool SafeQApplication::lastMessageWasSuppressed() noexcept
 void SafeQApplication::DebugOutput(
     QtMsgType type, const QMessageLogContext& context, const QString& msg)
 {
+#if defined(__EMSCRIPTEN__)
+  score::wasm::logMessage(type, context, msg);
+  if(type == QtFatalMsg)
+  {
+    SCORE_BREAKPOINT;
+    std::terminate();
+  }
+  return;
+#else
   auto basename_arr = QFileInfo(context.file).baseName().toUtf8();
   auto basename = basename_arr.constData();
   FILE* out_file = stderr;
 #if defined(_MSC_VER)
   static LogFile logger;
   out_file = logger.desc();
-#endif
-
-#if defined(__EMSCRIPTEN__)
-  // Never throttle a fatal: it is the last thing that will ever be printed.
-  if(type != QtFatalMsg)
-  {
-    bool suppress = false;
-    const qint64 repeats = throttleStep(msg, suppress);
-
-    if(repeats > 0)
-    {
-      const auto summary
-          = QStringLiteral("Info: [previous message repeated %1 times]").arg(repeats);
-      emitLine(QtInfoMsg, summary.toUtf8().constData());
-    }
-
-    t_lastSuppressed = suppress;
-    if(suppress)
-      return;
-  }
-  else
-  {
-    t_lastSuppressed = false;
-  }
-
-  {
-    static const char* const prefixes[]
-        = {"Debug", "Warning", "Critical", "Fatal", "Info"};
-    // QtDebugMsg=0, QtWarningMsg=1, QtCriticalMsg=2, QtFatalMsg=3, QtInfoMsg=4
-    const int idx = int(type);
-    const char* prefix
-        = (idx >= 0 && idx <= int(QtInfoMsg)) ? prefixes[idx] : "Log";
-    const QByteArray line = QStringLiteral("%1: %2 (%3:%4)")
-                                .arg(
-                                    QString::fromUtf8(prefix), msg,
-                                    QString::fromUtf8(basename))
-                                .arg(context.line)
-                                .toUtf8();
-    emitLine(type, line.constData());
-
-    if(type == QtFatalMsg)
-    {
-      SCORE_BREAKPOINT;
-      std::terminate();
-    }
-    return;
-  }
 #endif
 
   QByteArray localMsg = msg.toLocal8Bit();
@@ -198,6 +85,7 @@ void SafeQApplication::DebugOutput(
       std::terminate();
   }
   fflush(out_file);
+#endif
 }
 
 Q_GLOBAL_STATIC(QUrl, g_next_help_url_to_open);
